Fixed drop event duplicating local content.
If content is dropped into the editor from an external window, everything works but when done so from local content, images are duplicated and text will not move.

This commit forces the browser to perform its default action when actioning local content (and also skipping unnecessary compression on an already compressed item).

Signed-off-by: Finn Rayment <finn@rayment.fr>